+++ apps/designer/app/routes/rest/project.$projectId.tsx
@@ -1,4 +1,4 @@
-import { ActionFunction, json } from "@remix-run/node";
+import { LoaderFunction, json } from "@remix-run/node";
 import env from "~/env.server";
 import { db } from "@webstudio-is/project/server";
 import { sentryException } from "~/shared/sentry";
@@ -7,14 +7,14 @@ import type { Props, Tree } from "@webstudio-is/react-sdk";
 import type { Page } from "@webstudio-is/project";
 
 type PagesDetails = {
-  [key: string]: {
+  [path: string]: {
     page: Page;
     tree: Tree | null;
     props: Props[];
     css: string;
   };
 };
-export const loader: ActionFunction = async ({ request, params }) => {
+export const loader: LoaderFunction = async ({ request, params }) => {
   try {
     const projectId = params.projectId ?? undefined;
     const url = new URL(request.url);
@@ -27,7 +27,7 @@ export const loader: ActionFunction = async ({ request, params }) => {
     const prodBuild = await db.build.loadByProjectId(projectId, "prod");
     if (prodBuild === undefined) {
       throw json(
-        "Project not found or not published yet. Please contact us to get help.",
+        `Project ${projectId} not found or not published yet. Please contact us to get help.`,
         { status: 500 }
       );
     }
@@ -37,9 +37,12 @@ export const loader: ActionFunction = async ({ request, params }) => {
     const breakpoints = await db.breakpoints.load(prodBuild.id);
     const tree = await db.tree.loadById(homePage.treeId);
     if (tree === null) {
-      throw json("Tree structure not found. Please contact us to get help.", {
-        status: 500,
-      });
+      throw json(
+        `Tree ${homePage.treeId} structure not found. Please contact us to get help.`,
+        {
+          status: 500,
+        }
+      );
     }
     pagesDetails["/"] = {
       page: homePage,
@@ -48,16 +51,25 @@ export const loader: ActionFunction = async ({ request, params }) => {
       css: await generateCssText({
         projectId,
         mode: "published",
-        pathname: url.pathname,
+        pathname: "/",
         pageId: homePage.id,
       }),
     };
 
     if (pages.length > 0) {
       for (const page of pages) {
+        const tree = await db.tree.loadById(page.treeId);
+        if (tree === null) {
+          throw json(
+            `Tree ${page.treeId} structure not found. Please contact us to get help.`,
+            {
+              status: 500,
+            }
+          );
+        }
         pagesDetails[page.path] = {
           page,
-          tree: await db.tree.loadById(page.treeId),
+          tree,
           props: await db.props.loadByTreeId(page.treeId),
           css: await generateCssText({
             projectId,
