+++ packages/http-client/src/index.test.ts
@@ -1,5 +1,4 @@
 import { loadProject } from "./index";
-
 const existingProjectId = "675e8af3-48fa-4b18-9ebf-fd2b128865e2";
 const notPublishedProjectId = "7ec397c6-b3d0-4967-9073-9d83623fcf8e";
 const onlyHomeProjectId = "36d6c16f-04a0-45d4-ab1d-aa0ab61eb5b6";
@@ -16,7 +15,10 @@ describe("getProjectDetails", () => {
     if (response instanceof Error) {
       throw response;
     }
-    expect(Object.keys(response.pages).length > 1).toBeTruthy();
+    if (typeof response === "object") {
+      return expect(Object.keys(response.pages).length > 1).toBeTruthy();
+    }
+    throw new Error("Unexpected response");
   });
   test("does not include pages", async () => {
     const response = await loadProject({
@@ -26,7 +28,10 @@ describe("getProjectDetails", () => {
     if (response instanceof Error) {
       throw response;
     }
-    expect(Object.keys(response.pages).length === 1).toBeTruthy();
+    if (typeof response === "object") {
+      return expect(Object.keys(response.pages).length === 1).toBeTruthy();
+    }
+    throw new Error("Unexpected response");
   });
   test("loads existing project", async () => {
     const response = await loadProject({
@@ -41,7 +46,7 @@ describe("getProjectDetails", () => {
       projectId: notPublishedProjectId,
     });
     expect(response.toString()).toBe(
-      "Error: Project not found or not published yet. Please contact us to get help."
+      `Project ${notPublishedProjectId} not found or not published yet. Please contact us to get help.`
     );
   });
 });
