+++ packages/http-client/src/index.ts
@@ -1,5 +1,22 @@
 import fetch from "isomorphic-fetch";
-import type { Project } from "./index.d";
+import type { BaseInstance, UserProp } from "@webstudio-is/react-sdk";
+import type { Breakpoint } from "@webstudio-is/css-data";
+
+export type Page = {
+  id: string;
+  root: [string, string];
+};
+export type Project = {
+  pages: {
+    [path: string]: {
+      page: Page;
+      tree: BaseInstance;
+      props: Array<UserProp> | [];
+      breakpoints: Array<Breakpoint> | null;
+      css: string;
+    };
+  };
+};
 
 export const loadProject = async ({
   apiUrl,
@@ -7,7 +24,7 @@ export const loadProject = async ({
 }: {
   apiUrl: string;
   projectId: string;
-}): Promise<Project> => {
+}): Promise<Project | string> => {
   try {
     if (apiUrl === undefined) {
       throw new Error("Webstudio API URL is required.");
@@ -20,10 +37,10 @@ export const loadProject = async ({
       throw new Error(project);
     }
     return project;
-  } catch (error: unknown) {
+  } catch (error) {
     if (error instanceof Error) {
-      return error;
+      return error.message;
     }
-    return new Error("Unknown error");
+    throw error;
   }
 };
